Trait pallet_contracts_rpc::ContractsApiServer

pub trait ContractsApiServer<BlockHash, BlockNumber, AccountId, Balance, Hash>: Sized + Send + Sync + 'staticwhere
    Balance: Copy + TryFrom<NumberOrHex> + Into<NumberOrHex>,{
    fn call(
        &self,
        call_request: CallRequest<AccountId>,
        at: Option<BlockHash>
    ) -> RpcResult<ContractExecResult<Balance>>;
    fn instantiate(
        &self,
        instantiate_request: InstantiateRequest<AccountId, Hash>,
        at: Option<BlockHash>
    ) -> RpcResult<ContractInstantiateResult<AccountId, Balance>>;
    fn upload_code(
        &self,
        upload_request: CodeUploadRequest<AccountId>,
        at: Option<BlockHash>
    ) -> RpcResult<CodeUploadResult<Hash, Balance>>;
    fn get_storage(
        &self,
        address: AccountId,
        key: H256,
        at: Option<BlockHash>
    ) -> RpcResult<Option<Bytes>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        BlockNumber: Send + Sync + 'static,
        AccountId: Send + Sync + 'static + DeserializeOwned + Serialize,
        Balance: Send + Sync + 'static + Serialize + Copy + TryFrom<NumberOrHex> + Into<NumberOrHex>,
        Hash: Send + Sync + 'static + DeserializeOwned + Serialize,
    { ... }
}

Server trait implementation for the ContractsApi RPC API.

Required Methods

fn call(
&self,
call_request: CallRequest<AccountId>,
at: Option<BlockHash>
) -> RpcResult<ContractExecResult<Balance>>

Executes a call to a contract.

This call is performed locally without submitting any transactions. Thus executing this won’t change any state. Nonetheless, the calling state-changing contracts is still possible.

This method is useful for calling getter-like methods on contracts or to dry-run a a contract call in order to determine the gas_limit.

fn instantiate(
&self,
instantiate_request: InstantiateRequest<AccountId, Hash>,
at: Option<BlockHash>
) -> RpcResult<ContractInstantiateResult<AccountId, Balance>>

Instantiate a new contract.

This instantiate is performed locally without submitting any transactions. Thus the contract is not actually created.

This method is useful for UIs to dry-run contract instantiations.

fn upload_code(
&self,
upload_request: CodeUploadRequest<AccountId>,
at: Option<BlockHash>
) -> RpcResult<CodeUploadResult<Hash, Balance>>

Upload new code without instantiating a contract from it.

This upload is performed locally without submitting any transactions. Thus executing this won’t change any state.

This method is useful for UIs to dry-run code upload.

fn get_storage(
&self,
address: AccountId,
key: H256,
at: Option<BlockHash>
) -> RpcResult<Option<Bytes>>

Returns the value under a specified storage key in a contract given by address param, or None if it is not set.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where
BlockHash: Send + Sync + 'static + DeserializeOwned,
BlockNumber: Send + Sync + 'static,
AccountId: Send + Sync + 'static + DeserializeOwned + Serialize,
Balance: Send + Sync + 'static + Serialize + Copy + TryFrom<NumberOrHex> + Into<NumberOrHex>,
Hash: Send + Sync + 'static + DeserializeOwned + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.