Trait pallet_contracts_rpc::ContractsApi

pub trait ContractsApi<BlockHash, BlockNumber, AccountId, Balance, Hash>: Sized + Send + Sync + 'static where
    Balance: Copy + TryFrom<NumberOrHex> + Into<NumberOrHex>, {
    fn call(
        &self, 
        call_request: CallRequest<AccountId>, 
        at: Option<BlockHash>
    ) -> Result<ContractExecResult<Balance>>;

    fn instantiate(
        &self, 
        instantiate_request: InstantiateRequest<AccountId, Hash>, 
        at: Option<BlockHash>
    ) -> Result<ContractInstantiateResult<AccountId, Balance>>;

    fn upload_code(
        &self, 
        upload_request: CodeUploadRequest<AccountId>, 
        at: Option<BlockHash>
    ) -> Result<CodeUploadResult<Hash, Balance>>;

    fn get_storage(
        &self, 
        address: AccountId, 
        key: H256, 
        at: Option<BlockHash>
    ) -> Result<Option<Bytes>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        BlockNumber: Send + Sync + 'static,
        AccountId: Send + Sync + 'static + Serialize + DeserializeOwned,
        Balance: Send + Sync + 'static + Serialize + Copy + TryFrom<NumberOrHex> + Into<NumberOrHex>,
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
    { ... }
}

Contracts RPC methods.

Required methods

fn call(
&self,
call_request: CallRequest<AccountId>,
at: Option<BlockHash>
) -> Result<ContractExecResult<Balance>>

Executes a call to a contract.

This call is performed locally without submitting any transactions. Thus executing this won’t change any state. Nonetheless, the calling state-changing contracts is still possible.

This method is useful for calling getter-like methods on contracts or to dry-run a a contract call in order to determine the gas_limit.

fn instantiate(
&self,
instantiate_request: InstantiateRequest<AccountId, Hash>,
at: Option<BlockHash>
) -> Result<ContractInstantiateResult<AccountId, Balance>>

Instantiate a new contract.

This instantiate is performed locally without submitting any transactions. Thus the contract is not actually created.

This method is useful for UIs to dry-run contract instantiations.

fn upload_code(
&self,
upload_request: CodeUploadRequest<AccountId>,
at: Option<BlockHash>
) -> Result<CodeUploadResult<Hash, Balance>>

Upload new code without instantiating a contract from it.

This upload is performed locally without submitting any transactions. Thus executing this won’t change any state.

This method is useful for UIs to dry-run code upload.

fn get_storage(
&self,
address: AccountId,
key: H256,
at: Option<BlockHash>
) -> Result<Option<Bytes>>

Returns the value under a specified storage key in a contract given by address param, or None if it is not set.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> where
BlockHash: Send + Sync + 'static + DeserializeOwned,
BlockNumber: Send + Sync + 'static,
AccountId: Send + Sync + 'static + Serialize + DeserializeOwned,
Balance: Send + Sync + 'static + Serialize + Copy + TryFrom<NumberOrHex> + Into<NumberOrHex>,
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.