Trait pallet_contracts_rpc::ContractsApiClient

pub trait ContractsApiClient<BlockHash, BlockNumber, AccountId, Balance, Hash>: ClientT where
    BlockHash: Send + Sync + 'static + Serialize,
    BlockNumber: Send + Sync + 'static,
    AccountId: Send + Sync + 'static + Serialize + DeserializeOwned,
    Balance: Send + Sync + 'static + DeserializeOwned + Copy + TryFrom<NumberOrHex> + Into<NumberOrHex>,
    Hash: Send + Sync + 'static + Serialize + DeserializeOwned, {
    fn call<'life0, 'async_trait>(
        &'life0 self, 
        call_request: CallRequest<AccountId>, 
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ContractExecResult<Balance>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn instantiate<'life0, 'async_trait>(
        &'life0 self, 
        instantiate_request: InstantiateRequest<AccountId, Hash>, 
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ContractInstantiateResult<AccountId, Balance>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn upload_code<'life0, 'async_trait>(
        &'life0 self, 
        upload_request: CodeUploadRequest<AccountId>, 
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<CodeUploadResult<Hash, Balance>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn get_storage<'life0, 'async_trait>(
        &'life0 self, 
        address: AccountId, 
        key: H256, 
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Client implementation for the ContractsApi RPC API.

Provided methods

fn call<'life0, 'async_trait>(
&'life0 self,
call_request: CallRequest<AccountId>,
at: Option<BlockHash>
) -> Pin<Box<dyn Future<Output = RpcResult<ContractExecResult<Balance>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Executes a call to a contract.

This call is performed locally without submitting any transactions. Thus executing this won’t change any state. Nonetheless, the calling state-changing contracts is still possible.

This method is useful for calling getter-like methods on contracts or to dry-run a a contract call in order to determine the gas_limit.

fn instantiate<'life0, 'async_trait>(
&'life0 self,
instantiate_request: InstantiateRequest<AccountId, Hash>,
at: Option<BlockHash>
) -> Pin<Box<dyn Future<Output = RpcResult<ContractInstantiateResult<AccountId, Balance>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Instantiate a new contract.

This instantiate is performed locally without submitting any transactions. Thus the contract is not actually created.

This method is useful for UIs to dry-run contract instantiations.

fn upload_code<'life0, 'async_trait>(
&'life0 self,
upload_request: CodeUploadRequest<AccountId>,
at: Option<BlockHash>
) -> Pin<Box<dyn Future<Output = RpcResult<CodeUploadResult<Hash, Balance>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Upload new code without instantiating a contract from it.

This upload is performed locally without submitting any transactions. Thus executing this won’t change any state.

This method is useful for UIs to dry-run code upload.

fn get_storage<'life0, 'async_trait>(
&'life0 self,
address: AccountId,
key: H256,
at: Option<BlockHash>
) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Returns the value under a specified storage key in a contract given by address param, or None if it is not set.